app/requirements.txt
requests

Flask==3.0.3
Flask-Caching==2.3.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
alembic==1.13.1
gunicorn==22.0.0
watchdog==4.0.0
Werkzeug==3.0.3
WTForms==3.1.2
email-validator==2.1.1
PyJWT==2.8.0

trafilatura==1.9.0
beautifulsoup4==4.12.3
psycopg2==2.9.9
celery==5.4.0
redis==5.0.4

langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-core==0.1.52
onnxruntime==1.17.3
chromadb==0.5.0
